changed match results back to normal
